Enabled auto-incrementing of spruce formula via CI
The spruce CI pipeline needs to be able to give this
formula the shasum of what it will download, prior to having
created the github release. To make that (and install times of the formula)
easier, switched to downloading the raw binary, and added stub
comments for making sed less risky.